requirements.txt
google-api-python-client==2.100.0
google-auth==2.23.0
anthropic>=0.84.0
python-telegram-bot==21.0
python-dotenv==1.0.0
requests==2.31.0
